refactor(trim-paths): extract remap into own module
